feat(memory): complete Phase 4 streaming optimizations + integer overflow fixes   - Memory: 6 optimizations (2.5GB saved on 500K chains, 63% reduction).   - Streaming: exportChain() with Jackson JsonGenerator (constant 50MB).   - Integer overflow: int→long for block counts (10 classes).   - Tests: Fix streaming method signatures (2 tests).   - Docs: 3 memory reports (2059 lines), API updates.

+++ src/main/java/com/rbatllet/blockchain/dto/ChainExportData.java
@@ -18,7 +18,7 @@ public class ChainExportData {
     private EncryptionExportData encryptionData;
     private LocalDateTime exportTimestamp;
     private String version;
-    private int totalBlocks;
+    private long totalBlocks;
     private String description;
     private boolean hasEncryptedData;
     
@@ -76,11 +76,11 @@ public class ChainExportData {
         this.version = version;
     }
     
-    public int getTotalBlocks() {
+    public long getTotalBlocks() {
         return totalBlocks;
     }
     
-    public void setTotalBlocks(int totalBlocks) {
+    public void setTotalBlocks(long totalBlocks) {
         this.totalBlocks = totalBlocks;
     }
